Update font package version and add new font files
Bumped the version number of the custom fonts package from v1.3 to v1.4
in `default.nix` to reflect the addition of new fonts. Introduced two
new font files, `FairfaxPona.ttf` and `sitelen-pona-pona.otf`, under the
`src` directory. These additions enhance the variety of available fonts
in the project, responding to the requirement for expanded typographical
options in the application.